Split posts exporter tests into two separate files
One for the CSVTransform itself, one for the serializer
